meme-sniper-bot/requirements.txt
telethon>=1.36.0
aiohttp>=3.9.0
pyyaml>=6.0
solders>=0.21.0
solana>=0.34.0
base58>=2.1.0
anthropic>=0.39.0
anchorpy>=0.20.0
spl-token>=0.4.0
